Merge Form API & FormField API to Main API Branch

+++ Harris Pharmacy/HarrisPharamacy/HarrisPharmacy.Data/Services/FormService.cs
@@ -57,7 +57,7 @@ namespace HarrisPharmacy.Data.Services
         {
             return await _applicationDbContext.Forms
                 .Include(f => f.FormWithFields)
-                    .ThenInclude(f=>f.FormField)
+                    .ThenInclude(f => f.FormField)
                 .ToListAsync();
         }
 
@@ -138,15 +138,20 @@ namespace HarrisPharmacy.Data.Services
         /// <returns></returns>
         public async Task<Form> UpdateFormAsync(Form f, List<FormField> selectedFormFields)
         {
-            var form = await _applicationDbContext.Forms.FindAsync(f.FormId);
+            var form = await _applicationDbContext.Forms
+                .Include(f => f.FormWithFields)
+                .FirstOrDefaultAsync(form => form.FormId == f.FormId);
             if (form == null)
                 return null;
 
-            _applicationDbContext.FormWithFields.RemoveRange(form.FormWithFields);
+            if (form.FormWithFields != null)
+                _applicationDbContext.FormWithFields.RemoveRange(form.FormWithFields);
 
             List<FormWithFields> formWithFields = CreateFormWithFields(f.Description, selectedFormFields, form);
 
             form.FormWithFields = formWithFields;
+            form.Name = f.Name;
+            form.Description = f.Description;
             form.DateUpdated = DateTime.Now;
             _applicationDbContext.Forms.Update(form);
             await _applicationDbContext.SaveChangesAsync();
@@ -281,7 +286,7 @@ namespace HarrisPharmacy.Data.Services
                     DateUpdated = DateTime.Now,
                     Description = description,
                     Form = form,
-                    FormField = formField,
+                    FormField = null,
                     FormId = form.FormId,
                     FormFieldId = formField.FormFieldId
                 };
